Move AddMemberCommand and RemoveMemberCommand Validators to their own files.
